Automated replays <Florian | Filip>
Important notice: there is a bug in bwapi, due to which we always have to run these replays with an odd number of replays in the folder. For more information on the issue, see issue 497.
http://code.google.com/p/bwapi/issues/detail?id=497

(tried to solve it, but failed at compiling the old bwapi-version. It is a small work-around, which I find to be decent enough)